requirements.txt
asgiref==3.9.1
boto3==1.40.35
botocore==1.40.35
certifi==2025.8.3
charset-normalizer==3.4.3
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.6
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-import-export==4.3.9
django-storages==1.14.6
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
xlrd==2.0.2
